# Use GetPixelRowSpan to access pixel data in Histogram equalization