@brianpopow 2020-11-16
after the changes of this PR:

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.630 (2004/?/20H1)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100
  [Host]     : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  Job-ZJWYUQ : .NET Framework 4.8 (4.8.4250.0), X64 RyuJIT
  Job-PDUVYP : .NET Core 2.1.23 (CoreCLR 4.6.29321.03, CoreFX 4.6.29321.01), X64 RyuJIT
  Job-ZXDATS : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT

IterationCount=3  LaunchCount=1  WarmupCount=3

|                                               Method |        Job |       Runtime |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------------------- |----------- |-------------- |----------:|----------:|----------:|------:|------:|------:|----------:|
|                      'Global Histogram Equalization' | Job-YJFVOZ |    .NET 4.7.2 |  2.992 ms | 0.5856 ms | 0.0321 ms |     - |     - |     - |   5.22 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-YJFVOZ |    .NET 4.7.2 | 38.121 ms | 2.7466 ms | 0.1505 ms |     - |     - |     - |   6.77 KB |
|                      'Global Histogram Equalization' | Job-RDLXEE | .NET Core 2.1 |  2.299 ms | 0.6028 ms | 0.0330 ms |     - |     - |     - |   5.15 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-RDLXEE | .NET Core 2.1 | 18.264 ms | 2.2435 ms | 0.1230 ms |     - |     - |     - |   5.95 KB |
|                      'Global Histogram Equalization' | Job-VSRZSV | .NET Core 3.1 |  2.268 ms | 1.1815 ms | 0.0648 ms |     - |     - |     - |    6.9 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-VSRZSV | .NET Core 3.1 | 15.433 ms | 1.9711 ms | 0.1080 ms |     - |     - |     - |   5.95 KB |
```

before that:

```
|                                               Method |        Job |       Runtime |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------------------- |----------- |-------------- |----------:|----------:|----------:|------:|------:|------:|----------:|
|                      'Global Histogram Equalization' | Job-XFMRZJ |    .NET 4.7.2 |  3.061 ms | 1.3141 ms | 0.0720 ms |     - |     - |     - |   5.16 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-XFMRZJ |    .NET 4.7.2 | 48.401 ms | 6.5210 ms | 0.3574 ms |     - |     - |     - |    8.8 KB |
|                      'Global Histogram Equalization' | Job-WTJMNW | .NET Core 2.1 |  2.315 ms | 0.9744 ms | 0.0534 ms |     - |     - |     - |   5.13 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-WTJMNW | .NET Core 2.1 | 37.697 ms | 2.5760 ms | 0.1412 ms |     - |     - |     - |   5.95 KB |
|                      'Global Histogram Equalization' | Job-AVGHFY | .NET Core 3.1 |  2.287 ms | 0.6242 ms | 0.0342 ms |     - |     - |     - |   6.82 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-AVGHFY | .NET Core 3.1 | 16.601 ms | 3.7207 ms | 0.2039 ms |     - |     - |     - |   5.93 KB |
```

netcore 3.1 is almost the same as before, but net471 and core 2.1 are slower now.