@antonfirsov 2020-11-16
@brianpopow I guess before/after are the other way around in https://github.com/SixLabors/ImageSharp/pull/1431#issuecomment-728149117?

Any explanation for the significant difference for Framework VS Core? I would be surprised if it's only about indexers.