@brianpopow 2020-11-16
> @brianpopow I guess before/after are the other way around in [#1431 (comment)](https://github.com/SixLabors/ImageSharp/pull/1431#issuecomment-728149117)?
> 
> Any explanation for the significant difference for Framework VS Core? I would be surprised if it's only about indexers.

@antonfirsov  @JimBobSquarePants: I have run the benchmark now many times, but with the same results as above. The current master branch is actually **slower** than the change in this PR. I could not believe my eyes, maybe some else wants to double check this.

**After the change of this PR:**

```
|                                               Method |        Job |       Runtime |      Mean |    Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------------------- |----------- |-------------- |----------:|---------:|----------:|------:|------:|------:|----------:|
|                      'Global Histogram Equalization' | Job-KEXZPE |    .NET 4.7.2 |  3.028 ms | 1.898 ms | 0.1040 ms |     - |     - |     - |   5.16 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-KEXZPE |    .NET 4.7.2 | 38.288 ms | 3.331 ms | 0.1826 ms |     - |     - |     - |   7.43 KB |
|                      'Global Histogram Equalization' | Job-GHKMRJ | .NET Core 2.1 |  2.313 ms | 1.010 ms | 0.0553 ms |     - |     - |     - |   5.14 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-GHKMRJ | .NET Core 2.1 | 18.522 ms | 4.412 ms | 0.2419 ms |     - |     - |     - |   5.94 KB |
|                      'Global Histogram Equalization' | Job-DUVFTK | .NET Core 3.1 |  2.350 ms | 2.915 ms | 0.1598 ms |     - |     - |     - |    6.9 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-DUVFTK | .NET Core 3.1 | 15.641 ms | 4.962 ms | 0.2720 ms |     - |     - |     - |   5.91 KB |
```

**Current master branch:**

```
|                                               Method |        Job |       Runtime |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------------------- |----------- |-------------- |----------:|----------:|----------:|------:|------:|------:|----------:|
|                      'Global Histogram Equalization' | Job-WDJRLD |    .NET 4.7.2 |  3.008 ms | 0.5885 ms | 0.0323 ms |     - |     - |     - |   5.19 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-WDJRLD |    .NET 4.7.2 | 49.309 ms | 4.5882 ms | 0.2515 ms |     - |     - |     - |   7.88 KB |
|                      'Global Histogram Equalization' | Job-TYBBMO | .NET Core 2.1 |  2.249 ms | 0.7262 ms | 0.0398 ms |     - |     - |     - |   5.14 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-TYBBMO | .NET Core 2.1 | 37.419 ms | 2.8973 ms | 0.1588 ms |     - |     - |     - |   5.96 KB |
|                      'Global Histogram Equalization' | Job-OCZNHU | .NET Core 3.1 |  2.281 ms | 0.8250 ms | 0.0452 ms |     - |     - |     - |   6.89 KB |
| 'AdaptiveHistogramEqualization (Tile interpolation)' | Job-OCZNHU | .NET Core 3.1 | 16.639 ms | 4.2431 ms | 0.2326 ms |     - |     - |     - |   5.88 KB |
```

I cannot really tell where the large difference between the framework and core comes from.