@antonfirsov 2020-11-16
Can we add a test based on what we know about the root cause? (Maybe doesn't need the `20k x 10k` size.)