@brianpopow 2020-11-16
> Can we add a test based on what we know about the root cause? (Maybe doesn't need the `20k x 10k` size.)

It seems it has to be large to trigger this. 15k x 10k was the "smallest" image where i have seen this. This seems too large for a unit test in my opinion. @antonfirsov what do you think?